+++ SimPEG/EM/Static/IP/ProblemIP.py
@@ -89,7 +89,7 @@ class BaseIPProblem(BaseEMProblem):
                 dA_dmT = self.getADeriv(u_src, ATinvdf_duT, adjoint=True)
                 dRHS_dmT = self.getRHSDeriv(src, ATinvdf_duT, adjoint=True)
                 du_dmT = -dA_dmT + dRHS_dmT
-                Jtv += df_dmT + du_dmT
+                Jtv += (df_dmT + du_dmT).astype(float)
         # Conductivity ((d u / d log sigma).T)
         if self._formulation is 'EB':
             return -Utils.mkvc(Jtv)
